+++ VUA_api/Models/Subject.cs
@@ -49,16 +49,5 @@ namespace VUA_api
             hashCode = hashCode * 7 + isBUS.GetHashCode();
             return hashCode;
         }
-
-        public override string ToString()
-        {
-            string information;
-            if (isOptional) information = MainResources.Optional + "\r\n";
-            else information = MainResources.Mandatory + "\r\n";
-
-            if (isBUS) information = information + MainResources.BUS + "\r\n";
-            information = information + MainResources.SubjectName + base.ToString();
-            return information;
-        }
     }
 }
