+++ VUA_api/Properties/launchSettings.json
@@ -11,7 +11,7 @@
     "profiles": {
         "IIS Express": {
             "commandName": "IISExpress",
-            "launchBrowser": false,
+            "launchBrowser": true,
             "launchUrl": "api/Lecturer",
             "environmentVariables": {
                 "ASPNETCORE_ENVIRONMENT": "Development"
@@ -19,7 +19,7 @@
         },
         "VUA_api": {
             "commandName": "Project",
-            "launchBrowser": false,
+            "launchBrowser": true,
             "launchUrl": "Lecturer",
             "applicationUrl": "https://localhost:5001;http://localhost:5000",
             "environmentVariables": {
