+++ VUA_api/Scraper/log.txt
