+++ Vilnius-University-Advisor/DataNode.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Vilnius_University_Advisor
 {
-    public class DataNode : IEquatable<DataNode>, IComparable<DataNode>
+    public class DataNode
     {
         public string name { get; set; }
         public Faculty faculty { get; set; }
@@ -21,7 +21,6 @@ namespace Vilnius_University_Advisor
             }
         }
         public int numberOfReviews { get; set; }
-
         public List<Review> reviews;
         public DataNode(string name, Faculty faculty)
         {
@@ -30,33 +29,6 @@ namespace Vilnius_University_Advisor
             this.scoreField = 0;
             this.numberOfReviews = 0;
             this.reviews = new List<Review>();
-        }
-
-        public override bool Equals(object obj)
-        {
-            DataNode item = obj as DataNode;
-            if (item == null) return false;
-            if (name.Equals(item.name) && faculty == item.faculty) return true;
-            else return false;
-        }
-        public override int GetHashCode()
-        {
-            int hashCode = name.GetHashCode();
-            hashCode = hashCode * 7 + faculty.GetHashCode();
-            return hashCode;
-        }
-
-        public bool Equals(DataNode other)
-        {
-            return other.name.Equals(this.name) && other.faculty==this.faculty;
-        }
-
-        public int CompareTo(DataNode other)
-        {
-            int result = this.name.CompareTo(other.name);
-            if (result == 0)
-                return this.faculty.CompareTo(other.faculty);
-            else return result;
         }
         public override string ToString()
         {
