--- Vilnius-University-Advisor/Scraper/Fetcher.cs
@@ -1,41 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Net;
-using System.Text;
-using System.Threading;
-using System.Windows.Forms;
-
-namespace Vilnius_University_Advisor.Scraper
-{
-    class Fetcher
-    {
-        WebBrowser webBrowser1 = new WebBrowser();
-        public string GetTimetableText(String url)
-        {           
-            webBrowser1.ScriptErrorsSuppressed = true;
-            webBrowser1.Navigate(url);
-
-            while (webBrowser1.ReadyState != WebBrowserReadyState.Complete)
-            {
-                Application.DoEvents();
-            }
-            for(int i=0; i<500000; i++) Application.DoEvents();
-
-            var documentAsIHtmlDocument3 = (mshtml.IHTMLDocument3)webBrowser1.Document.DomDocument;
-            return documentAsIHtmlDocument3.documentElement.outerHTML;
-        }
-        public string GetFacultyHtml(String url)
-        {
-            webBrowser1.ScriptErrorsSuppressed = true;
-            webBrowser1.Navigate(url);
-
-            while (webBrowser1.ReadyState != WebBrowserReadyState.Complete)
-            {
-                Application.DoEvents();
-            }
-
-            var documentAsIHtmlDocument3 = (mshtml.IHTMLDocument3)webBrowser1.Document.DomDocument;
-            return documentAsIHtmlDocument3.documentElement.outerHTML;
-        }
-    }
-}
